requirements.txt
mmcls==0.15.0
torch==1.6.0
mmcv-full==1.3.13
torchvision==0.7.0
pydicom==2.1.2
scikit-learn==0.24.2
timm==0.4.12